MX-137: Implement the self-change-address finalizing mutation
