+++ jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
   preset: "ts-jest",
-  testEnvironment: "node"
+  testEnvironment: "node",
+  testPathIgnorePatterns: ["/node_modules/", "<rootDir>/dist/"]
 };
