+++ package.json
@@ -2,7 +2,7 @@
   "name": "giraffe",
   "main": "dist/index.js",
   "scripts": {
-    "test": "jest --silent",
+    "test": "jest",
     "start": "node .",
     "dev": "ts-node --transpile-only src/index.ts",
     "build": "tsc && cp src/schema.graphqls dist/schema.graphqls && mkdir -p dist/translations && cp -r src/translations/*.json dist/translations",
@@ -49,7 +49,7 @@
     "@types/dotenv": "^6.1.1",
     "@types/graphql": "^14.2.3",
     "@types/ioredis": "^4.0.17",
-    "@types/jest": "^24.0.23",
+    "@types/jest": "^26.0.23",
     "@types/koa": "^2.0.49",
     "@types/koa-compress": "^2.0.9",
     "@types/node-fetch": "^2.5.0",
@@ -60,11 +60,11 @@
     "apollo-server-testing": "2.21.0",
     "concurrently": "^4.1.1",
     "graphql-schema-typescript": "^1.5.2",
-    "jest": "^24.9.0",
+    "jest": "^26.6.3",
     "jest-expo": "^41.0.0",
     "nodemon": "^1.19.1",
     "prettier": "^1.18.2",
-    "ts-jest": "^24.2.0",
+    "ts-jest": "^26.5.6",
     "ts-node": "^8.0.3",
     "tslint": "^5.18.0",
     "tslint-config-prettier": "^1.18.0",
