+++ src/api/upstreams/underwriter.ts
@@ -1,8 +1,11 @@
-
 import { HttpClient } from '../httpClient'
 
 export interface UnderwriterClient {
   createQuote(body: CreateQuoteDto): Promise<QuoteCreationResult>
+  changeToQuotes(
+    contractIds: string[],
+    quoteIds: string[],
+  ): Promise<QuoteChangingResult>
 }
 
 export interface CreateQuoteDto {
@@ -27,7 +30,6 @@ export interface CreateQuoteDto {
     ancillaryArea: number
     yearOfConstruction: number
     numberOfBathrooms: number
-    floor: number
     subleted: boolean
     extraBuildings: {
       type: string
@@ -69,17 +71,19 @@ export interface CreateQuoteDto {
   }
 }
 
-export type QuoteCreationResult = QuoteCreationSuccessDto | QuoteCreationFailureDto
+export type QuoteCreationResult =
+  | QuoteCreationSuccessDto
+  | QuoteCreationFailureDto
 
 export interface QuoteCreationSuccessDto {
-  status: 'success',
+  status: 'success'
   id: string
   price: number
   validTo: string
 }
 
 export interface QuoteCreationFailureDto {
-  status: 'failure',
+  status: 'failure'
   breachedUnderwritingGuidelines: BreachedGuidelineDto[]
 }
 
@@ -87,16 +91,31 @@ export interface BreachedGuidelineDto {
   code: string
 }
 
-export const createUnderwriterClient = (client: HttpClient): UnderwriterClient => {
+export interface QuoteChangingResult {
+  updatedContractIds: string[]
+  createdContractIds: string[]
+  terminatedContractIds: string[]
+}
+
+export const createUnderwriterClient = (
+  client: HttpClient,
+): UnderwriterClient => {
   return {
     createQuote: async (body) => {
       // 422 means underwriting guidelines breached, and we should read the body
-      const res = await client.post("/quotes", body, { validStatusCodes: [422] })
+      const res = await client.post('/quotes', body, {
+        validStatusCodes: [422],
+      })
       const result = await res.json()
       if (res.status === 422) {
-        return { status: 'failure', ...result}
+        return { status: 'failure', ...result }
       }
-      return { status: 'success', ...result}
-    }
+      return { status: 'success', ...result }
+    },
+    changeToQuotes: async (contractIds, quoteIds) => {
+      const body = { contractIds, quoteIds }
+      const result = await client.post('/quotes/selfChange', body)
+      return await result.json()
+    },
   }
 }
