+++ src/resolvers/index.ts
@@ -68,7 +68,7 @@ import { geo } from './geo'
 import { startDirectDebitRegistration } from './trustly'
 import { uploadFile, uploadFiles } from './uploadFile'
 import { activeContractBundles, contracts, hasContract } from './contracts'
-import { selfChangeEligibility, createAddressChangeQuotes } from './selfChange'
+import { createAddressChangeQuotes, commitAddressChangeQuotes } from './selfChange'
 import { availableLocales } from './availableLocales'
 
 const resolvers: Resolver = {
@@ -90,7 +90,6 @@ const resolvers: Resolver = {
     activeContractBundles,
     contracts,
     hasContract,
-    selfChangeEligibility,
     availableLocales
   },
   Mutation: {
@@ -127,7 +126,8 @@ const resolvers: Resolver = {
     registerBranchCampaign,
     updateLanguage,
     updatePickedLocale,
-    createAddressChangeQuotes
+    createAddressChangeQuotes,
+    commitAddressChangeQuotes
   },
   Subscription: {
     offer: subscribeToOffer,
