+++ src/resolvers/selfChange.ts
@@ -1,46 +1,21 @@
-import { QuoteCreationResult } from './../api/upstreams/underwriter';
-import { convertAddressChangeToSelfChangeBody } from './selfChangeTranslator'
+import { MemberDto } from './../api/upstreams/memberService';
+import { QuoteCreationResult, CreateQuoteDto } from './../api/upstreams/underwriter';
 import {
   MutationToCreateAddressChangeQuotesResolver,
   AddressChangeQuoteResult,
   PossibleAddressChangeQuoteResultTypeNames,
+  AddressChangeInput,
+  AddressOwnership,
+  AddressHomeType,
+  MutationToCommitAddressChangeQuotesResolver,
+  AddressChangeCommitResult,
 } from './../typings/generated-graphql-types'
 
-import {
-  QueryToSelfChangeEligibilityResolver,
-  SelfChangeEligibility,
-} from '../typings/generated-graphql-types'
-import { ContractStatusDto } from '../api/upstreams/productPricing'
+import { ContractDto, ContractMarketInfoDto, ContractStatusDto } from '../api/upstreams/productPricing'
 import { Typenamed } from '../utils/types';
+import { bundleContracts } from './helpers/contractTransformers';
 
-const ADDRESS_CHANGE_STORIES_BY_MARKET: Record<string, string> = {
-  SWEDEN: 'moving-flow-SE',
-  NORWAY: 'moving-flow-NO',
-}
-
-const INELIGIBLE: SelfChangeEligibility = {
-  blockers: [],
-}
-
-const selfChangeEligibility: QueryToSelfChangeEligibilityResolver = async (
-  _parent,
-  _args,
-  { upstream },
-): Promise<SelfChangeEligibility> => {
-  const eligibility = await upstream.productPricing.getSelfChangeEligibility()
-  const isEligible = eligibility.blockers.length == 0
-  if (!isEligible) {
-    return INELIGIBLE
-  }
-
-  const marketInfo = await upstream.productPricing.getContractMarketInfo()
-  return {
-    blockers: [], // is deprecated
-    addressChangeEmbarkStoryId: ADDRESS_CHANGE_STORIES_BY_MARKET[marketInfo.market.toUpperCase()],
-  }
-}
-
-const createAddressChangeQuotes: MutationToCreateAddressChangeQuotesResolver = async (
+export const createAddressChangeQuotes: MutationToCreateAddressChangeQuotesResolver = async (
   _parent,
   args,
   { upstream },
@@ -65,6 +40,185 @@ const createAddressChangeQuotes: MutationToCreateAddressChangeQuotesResolver = a
 
   const responses = await Promise.all(tasks)
   return transformResult(responses)
+}
+
+export const commitAddressChangeQuotes: MutationToCommitAddressChangeQuotesResolver = async (
+  _parent,
+  args,
+  { upstream, strings }
+): Promise<AddressChangeCommitResult> => {
+  const contractIds = args.contractBundleId.replace('bundle:', '').split(',')
+  const result = await upstream.underwriter.changeToQuotes(contractIds, args.quoteIds)
+  const [
+    createdContracts,
+    updatedContracts
+  ] = await Promise.all([
+    Promise.all(result.createdContractIds.map(upstream.productPricing.getContract)),
+    Promise.all(result.updatedContractIds.map(upstream.productPricing.getContract))
+  ])
+
+  const bundle = bundleContracts(strings, createdContracts.concat(updatedContracts))
+  if (bundle.length !== 1) {
+    throw `Unexpected bundle size after self changed address, ids: ${bundle.map(b => b.id)}`
+  }
+  return {
+    newContractBundle: bundle[0]
+  }
+}
+
+const convertAddressChangeToSelfChangeBody = (
+  input: AddressChangeInput,
+  member: MemberDto,
+  contract: ContractDto,
+  marketInfo: ContractMarketInfoDto
+): CreateQuoteDto => {
+  const dtoBase: CreateQuoteDto = {
+    memberId: member.memberId,
+    firstName: member.firstName,
+    lastName: member.lastName,
+    ssn: member.ssn,
+    birthDate: member.birthDate,
+    startDate: input.startDate,
+  }
+
+  switch (marketInfo.market) {
+    case 'SWEDEN':
+      return toSwedishQuoteDto(input, dtoBase)
+    case 'NORWAY':
+      return toNorwegianQuoteDto(input, dtoBase, contract)
+    case 'DENMARK':
+      return toDanishQuoteDto(input, dtoBase, contract)
+  }
+
+  throw `Unhandled market: ${marketInfo.market}`
+}
+
+const toSwedishQuoteDto = (
+  input: AddressChangeInput,
+  dto: CreateQuoteDto,
+): CreateQuoteDto => {
+  const subtypeMapper = (
+    ownership: AddressOwnership,
+    isStudent: boolean,
+  ): string => {
+    switch (ownership) {
+      case AddressOwnership.OWN:
+        return isStudent ? 'BRF_STUDENT' : 'BRF'
+      case AddressOwnership.RENT:
+        return isStudent ? 'RENT_STUDENT' : 'RENT'
+    }
+  }
+
+  switch (input.type) {
+    case AddressHomeType.APARTMENT:
+      return {
+        ...dto,
+        swedishApartmentData: {
+          street: input.street,
+          zipCode: input.zip,
+          householdSize: input.numberCoInsured + 1,
+          livingSpace: input.livingSpace,
+          subType: subtypeMapper(input.ownership, input.isStudent!!),
+        },
+      }
+    case AddressHomeType.HOUSE:
+      return {
+        ...dto,
+        swedishHouseData: {
+          street: input.street,
+          zipCode: input.zip,
+          householdSize: input.numberCoInsured + 1,
+          livingSpace: input.livingSpace,
+          ancillaryArea: input.ancillaryArea!!,
+          yearOfConstruction: input.yearOfConstruction!!,
+          numberOfBathrooms: input.numberOfBathrooms!!,
+          subleted: input.isSubleted!!,
+          extraBuildings: input.extraBuildings!!,
+        },
+      }
+  }
+}
+
+const toNorwegianQuoteDto = (
+  input: AddressChangeInput,
+  dto: CreateQuoteDto,
+  contract: ContractDto,
+): CreateQuoteDto => {
+  const isHomeContent = contract.typeOfContract.startsWith('NO_HOME_CONTENT')
+  const isTravel = contract.typeOfContract.startsWith('NO_TRAVEL')
+
+  if (isHomeContent) {
+    return {
+      ...dto,
+      norwegianHomeContentsData: {
+        street: input.street,
+        zipCode: input.zip,
+        coInsured: input.numberCoInsured,
+        livingSpace: input.livingSpace,
+        subType: input.ownership,
+        youth: input.isYouth!,
+      },
+    }
+  }
+  if (isTravel) {
+    return {
+      ...dto,
+      norwegianTravelData: {
+        coInsured: input.numberCoInsured,
+        youth: input.isYouth!,
+      },
+    }
+  }
+
+  throw `Unhandled type of contract: ${contract.typeOfContract}`
+}
+
+const toDanishQuoteDto = (
+  input: AddressChangeInput,
+  dto: CreateQuoteDto,
+  contract: ContractDto,
+): CreateQuoteDto => {
+  const isHomeContent = contract.typeOfContract.startsWith('DK_HOME_CONTENT')
+  const isTravel = contract.typeOfContract.startsWith('DK_TRAVEL')
+  const isAccident = contract.typeOfContract.startsWith('DK_ACCIDENT')
+
+  if (isHomeContent) {
+    return {
+      ...dto,
+      danishHomeContentsData: {
+        street: input.street,
+        zipCode: input.zip,
+        coInsured: input.numberCoInsured,
+        livingSpace: input.livingSpace,
+        subType: input.ownership,
+        student: input.isStudent!,
+      },
+    }
+  }
+  if (isTravel) {
+    return {
+      ...dto,
+      danishTravelData: {
+        street: input.street,
+        zipCode: input.zip,
+        coInsured: input.numberCoInsured,
+        student: input.isStudent!,
+      },
+    }
+  }
+  if (isAccident) {
+    return {
+      ...dto,
+      danishAccidentData: {
+        street: input.street,
+        zipCode: input.zip,
+        coInsured: input.numberCoInsured,
+        student: input.isStudent!,
+      },
+    }
+  }
+
+  throw `Unhandled type of contract: ${contract.typeOfContract}`
 }
 
 const transformResult = (
@@ -93,5 +247,3 @@ const transformResult = (
     quoteIds: quoteIds
   }
 }
-
-export { selfChangeEligibility, createAddressChangeQuotes }
