+++ src/schema.graphqls
@@ -21,7 +21,7 @@ type Query {
   """Returns whether a member has at least one contract"""
   hasContract: Boolean!
   """Returns a type describing whether the 'Self Change' functionality is possible."""
-  selfChangeEligibility: SelfChangeEligibility!
+  selfChangeEligibility: SelfChangeEligibility! @deprecated(reason: "Use angelStories in `activeContractBundles` instead")
   """All locales that are available and activated"""
   availableLocales: [Locale!]!
 }
@@ -62,6 +62,8 @@ type Mutation {
   updatePickedLocale(pickedLocale: Locale!): Member!
   """Create all the quotes needed in relation to a change of address, based on the current state of the member's insurance."""
   createAddressChangeQuotes(input: AddressChangeInput!): AddressChangeQuoteResult!
+  """Commit the address change quotes, installing them as the ."""
+  commitAddressChangeQuotes(contractBundleId: ID!, quoteIds: [ID!]!): AddressChangeCommitResult!
 }
 
 enum Locale {
@@ -397,8 +399,6 @@ input AddressChangeInput {
   yearOfConstruction: Int
   """Number of bathrooms. Required if type == HOUSE."""
   numberOfBathrooms: Int
-  """Number of floors. Required if type == HOUSE."""
-  numberOfFloors: Int
   """Is this property subleted? Required if type == HOUSE."""
   isSubleted: Boolean
   """A list of extra buildings outside of the main property. Required if type == HOUSE."""
@@ -429,6 +429,10 @@ type AddressChangeQuoteSuccess {
 
 type AddressChangeQuoteFailure {
   breachedUnderwritingGuidelines: [String!]!
+}
+
+type AddressChangeCommitResult {
+  newContractBundle: ContractBundle!
 }
 
 enum LoggingSource {
